# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HydraGuard API
  version: 0.1.0
  description: >
    Evidence-first supply-chain ingestion and incident analysis API.
    Registry declarations, lockfile resolutions, and incident evidence
    remain separate so candidate exposure is not confused with confirmed
    package resolution.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
  /typosquatting/scans:
    post:
      tags: [Typosquatting]
      operationId: createTyposquattingScan
      summary: Scan an npm lockfile for typosquatting findings
      description: >
        Accepts the same evidence-backed lockfile request as lockfile ingestion
        and asynchronously scans its exact package resolutions for
        typosquatting candidates.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LockfileIngestionRequest"
      responses:
        "202":
          description: Typosquatting scan accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: A referenced resource was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          description: The scan could not be created due to an internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The scan service is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /typosquatting/findings:
    get:
      tags: [Typosquatting]
      operationId: listTyposquattingFindings
      summary: List typosquatting findings
      description: >
        Returns a bounded page of findings. cursorDetectedAt and
        cursorFindingId must either both be supplied or both be omitted.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursorDetectedAt
          in: query
          required: false
          description: Supply together with cursorFindingId.
          schema:
            type: integer
            minimum: 0
            maximum: 9007199254740991
        - name: cursorFindingId
          in: query
          required: false
          description: Supply together with cursorDetectedAt.
          schema:
            type: integer
            minimum: 0
            maximum: 9007199254740991
      responses:
        "200":
          description: A bounded page of typosquatting findings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TyposquattingFindingListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: A referenced resource was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          description: Findings could not be listed due to an internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The findings service is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /typosquatting/findings/{findingId}:
    get:
      tags: [Typosquatting]
      operationId: getTyposquattingFinding
      summary: Get a typosquatting finding
      parameters:
        - $ref: "#/components/parameters/TyposquattingFindingId"
      responses:
        "200":
          description: Detailed finding, evidence, versions, and exposure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TyposquattingFindingDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: The finding was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          description: The finding could not be read due to an internal error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The findings service is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    analystBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque analyst token
      description: The API maps this token to a trusted server-configured analyst principal.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Stable caller-generated key preventing duplicate asynchronous jobs.
      schema:
        type: string
        minLength: 8
        maxLength: 80
        pattern: "^[A-Za-z0-9._-]{8,80}$"
...
    TyposquattingFindingId:
      name: findingId
      in: path
      required: true
      schema:
        type: string
        pattern: "^(0|[1-9][0-9]{0,15})$"
        maxLength: 16
  responses:
...
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Identity or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    TyposquattingId:
      type: integer
      minimum: 0
      maximum: 9007199254740991
    TyposquattingFindingStatus:
      type: string
      enum:
        - candidate
        - suspicious
        - high-confidence
        - confirmed
        - dismissed
    TyposquattingTransformation:
      type: string
      enum:
        - adjacent-transposition
        - insertion
        - deletion
        - substitution
        - separator-variation
        - repeated-character
        - scope-impersonation
        - unicode-confusable
        - prefix-suffix
    TyposquattingFindingSummary:
      type: object
      additionalProperties: false
      required:
        - findingId
        - status
        - score
        - scoreMeaning
        - candidateName
        - targetName
        - summary
        - transformations
        - reasonCodes
        - detectedAt
        - synthetic
      properties:
        findingId:
          $ref: "#/components/schemas/TyposquattingId"
        status:
          $ref: "#/components/schemas/TyposquattingFindingStatus"
        score:
          type: number
          minimum: 0
          maximum: 100
        scoreMeaning:
          type: string
          enum: [heuristic-ranking-not-probability]
        candidateName:
          type: string
          minLength: 1
          maxLength: 214
        targetName:
          type: string
          minLength: 1
          maxLength: 214
        summary:
          type: string
          minLength: 1
          maxLength: 2000
        transformations:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TyposquattingTransformation"
        reasonCodes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 128
        detectedAt:
          $ref: "#/components/schemas/TyposquattingId"
        decidedAt:
          $ref: "#/components/schemas/TyposquattingId"
        decisionReason:
          type: string
          minLength: 1
          maxLength: 2000
        synthetic:
          type: boolean
    TyposquattingFindingCursor:
      type: object
      additionalProperties: false
      required:
        - detectedAt
        - findingId
      properties:
        detectedAt:
          $ref: "#/components/schemas/TyposquattingId"
        findingId:
          $ref: "#/components/schemas/TyposquattingId"
    TyposquattingFindingListResponse:
      type: object
      additionalProperties: false
      required:
        - findings
        - truncated
      properties:
        findings:
          type: array
          items:
            $ref: "#/components/schemas/TyposquattingFindingSummary"
        truncated:
          type: boolean
        nextCursor:
          $ref: "#/components/schemas/TyposquattingFindingCursor"
    TyposquattingEvidenceSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - sourceType
        - confidence
        - observedAt
        - synthetic
      properties:
        id:
          $ref: "#/components/schemas/TyposquattingId"
        sourceType:
          type: string
          minLength: 1
          maxLength: 64
        confidence:
          type: number
          minimum: 0
          maximum: 1
        observedAt:
          $ref: "#/components/schemas/TyposquattingId"
        synthetic:
          type: boolean
    TyposquattingExactVersion:
      type: object
      additionalProperties: false
      required:
        - id
        - version
        - synthetic
      properties:
        id:
          $ref: "#/components/schemas/TyposquattingId"
        version:
          type: string
          minLength: 1
          maxLength: 256
        synthetic:
          type: boolean
    TyposquattingVersionLookup:
      type: object
      additionalProperties: false
      required:
        - scannedCount
        - truncated
      properties:
        scannedCount:
          type: integer
          minimum: 0
          maximum: 1000
        truncated:
          type: boolean
    TyposquattingExposureService:
      type: object
      additionalProperties: false
      required:
        - serviceId
        - serviceLogicalId
        - serviceName
        - serviceCriticality
        - packageVersionIds
        - evidenceIds
      properties:
        serviceId:
          $ref: "#/components/schemas/TyposquattingId"
        serviceLogicalId:
          type: string
          minLength: 1
          maxLength: 512
        serviceName:
          type: string
          minLength: 1
          maxLength: 200
        serviceCriticality:
          type: string
          enum: [low, medium, high, critical]
        packageVersionIds:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TyposquattingId"
        evidenceIds:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TyposquattingId"
    TyposquattingExposureLimits:
      type: object
      additionalProperties: false
      required:
        - maxDepth
        - maxServices
        - maxTraversalStates
        - maxDependentsPerNode
      properties:
        maxDepth:
          type: integer
          minimum: 1
        maxServices:
          type: integer
          minimum: 1
        maxTraversalStates:
          type: integer
          minimum: 1
        maxDependentsPerNode:
          type: integer
          minimum: 1
    TyposquattingExposure:
      type: object
      additionalProperties: false
      required:
        - services
        - truncated
        - traversalStates
        - limits
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/TyposquattingExposureService"
        truncated:
          type: boolean
        traversalStates:
          type: integer
          minimum: 0
          maximum: 10000
        limits:
          $ref: "#/components/schemas/TyposquattingExposureLimits"
    TyposquattingFindingDetailResponse:
      type: object
      additionalProperties: false
      required:
        - finding
        - candidatePackageId
        - targetPackageId
        - evidence
        - exactVersions
        - versionLookup
        - exposure
        - incidentIds
      properties:
        finding:
          $ref: "#/components/schemas/TyposquattingFindingSummary"
        candidatePackageId:
          $ref: "#/components/schemas/TyposquattingId"
        targetPackageId:
          $ref: "#/components/schemas/TyposquattingId"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/TyposquattingEvidenceSummary"
        exactVersions:
          type: array
          items:
            $ref: "#/components/schemas/TyposquattingExactVersion"
        versionLookup:
          $ref: "#/components/schemas/TyposquattingVersionLookup"
        exposure:
          $ref: "#/components/schemas/TyposquattingExposure"
        incidentIds:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/TyposquattingId"
...
    ProvenanceInput:
      type: object
      additionalProperties: false
      required:
        - sourceType
        - sourceUri
        - observedAt
        - collectorVersion
        - confidence
        - synthetic
      properties:
        sourceType:
          type: string
          enum:
            - npm-registry
            - package-manifest
            - package-lock
            - git-commit
            - cyclonedx
            - spdx
            - slsa
            - sigstore
            - runtime-telemetry
            - security-advisory
            - synthetic-fixture
            - other
        sourceUri:
          type: string
          format: uri
        observedAt:
          type: string
          format: date-time
        collectorVersion:
          type: string
          minLength: 1
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        synthetic:
          type: boolean
...
    LockfileIngestionRequest:
      type: object
      additionalProperties: false
      required:
        - serviceLogicalId
        - serviceName
        - serviceCriticality
        - format
        - lockfile
        - provenance
      properties:
        serviceLogicalId:
          type: string
          pattern: "^service:"
          example: "service:demo-org:payment-api"
        repositoryLogicalId:
          type: string
          pattern: "^repo:"
          example: "repo:github:demo-org/payment-api"
        format:
          type: string
          enum:
            - npm-package-lock-v2
            - npm-package-lock-v3
        lockfile:
          type: object
          description: Parsed package-lock.json JSON object.
          additionalProperties: true
        provenance:
          $ref: "#/components/schemas/ProvenanceInput"
        serviceName:
          type: string
          minLength: 1

        serviceCriticality:
          type: string
          enum: [low, medium, high, critical]
...
    IngestionAccepted:
      type: object
      additionalProperties: false
      required:
        - ingestionId
        - status
        - submittedAt
      properties:
        ingestionId:
          type: string
        status:
          type: string
          enum: [queued]
        submittedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
...